requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==1.11.7
django-bootstrap4==0.0.7
django-ckeditor==5.6.1
django-filter==2.0.0
django-import-export==1.1.0
django-jet==1.0.8
django-js-asset==1.1.0
django-markdownx==2.0.28
django-multi-form-view==2.0.0
django-pagedown==1.0.6
django-widget-tweaks==1.4.3
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
Markdown==3.0.1
mpmath==1.0.0
numpy==1.15.4
odfpy==1.4.0
openpyxl==2.5.12
pandas==0.23.4
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
six==1.11.0
sorl-thumbnail==12.5.0
sympy==1.3
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.1.0
xlwt==1.3.0
